Crate fastsim_cli

Structs

AdoptHDResults ğŸ”’

AdoptResults ğŸ”’

ArrayObject ğŸ”’

FastSimApi ğŸ”’

Wrapper for fastsim. After running cargo build --release, run with

H2AndDieselResults

ParsedValue ğŸ”’

Functions

array_to_object_representation ğŸ”’

Takes a vector of floats and transforms it into an object representation used by the ndarray library.

calculate_mpgge_for_h2_diesel_ice

integrate_power_to_kwh

json_rewrite ğŸ”’

Rewrites the ADOPT JSON string to be in compliance with what FASTSim expects for JSON input.

main

transform_array_of_value_to_ndarray_representation ğŸ”’

transform_array_of_value_to_vec_of_f64 ğŸ”’

translate_veh_pt_type ğŸ”’